cxx-qt-gen: reduce generation tests as we don't pattern match anymore
Now that we don't pattern match on types anymore, we can reduce
the amount of tests we have for the expected output for a given
input. Now we just have tests for properties, invokables, signals,
and general passthrough or naming.
