+++ crates/cxx-qt-gen/test_inputs/passthrough_and_naming.rs.license
