+++ crates/cxx-qt-gen/test_outputs/passthrough_and_naming.rs.license
