--- crates/cxx-qt-gen/test_outputs/types_qt_invokable.cpp
@@ -1,153 +1,0 @@
-#include "cxx-qt-gen/include/my_object.cxxqt.h"
-
-namespace cxx_qt::my_object {
-
-MyObject::MyObject(QObject* parent)
-  : QObject(parent)
-  , m_rustObj(cxx_qt::my_object::cxx_qt_my_object::createRs())
-  , m_rustObjMutex(std::make_shared<std::mutex>())
-  , m_cxxQtThreadObj(
-      std::make_shared<rust::cxxqtlib1::CxxQtGuardedPointer<MyObject>>(this))
-{
-}
-
-MyObject::~MyObject()
-{
-  const auto guard = std::unique_lock(m_cxxQtThreadObj->mutex);
-  m_cxxQtThreadObj->ptr = nullptr;
-}
-
-const MyObjectRust&
-MyObject::unsafeRust() const
-{
-  return *m_rustObj;
-}
-
-MyObjectRust&
-MyObject::unsafeRustMut()
-{
-  return *m_rustObj;
-}
-
-std::unique_ptr<MyObjectCxxQtThread>
-MyObject::qtThread() const
-{
-  return std::make_unique<MyObjectCxxQtThread>(m_cxxQtThreadObj,
-                                               m_rustObjMutex);
-}
-
-QColor
-MyObject::testColor(const QColor& color)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QColor, ::std::unique_ptr<QColor>>{}(
-    m_rustObj->testColorWrapper(*this, color));
-}
-
-QDate
-MyObject::testDate(const QDate& date)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QDate, QDate>{}(
-    m_rustObj->testDateWrapper(*this, date));
-}
-
-QDateTime
-MyObject::testDateTime(const QDateTime& dateTime)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QDateTime,
-                                         ::std::unique_ptr<QDateTime>>{}(
-    m_rustObj->testDateTimeWrapper(*this, dateTime));
-}
-
-QPoint
-MyObject::testPoint(const QPoint& point)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QPoint, QPoint>{}(
-    m_rustObj->testPointWrapper(*this, point));
-}
-
-QPointF
-MyObject::testPointf(const QPointF& pointf)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QPointF, QPointF>{}(
-    m_rustObj->testPointfWrapper(*this, pointf));
-}
-
-QRect
-MyObject::testRect(const QRect& rect)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QRect, QRect>{}(
-    m_rustObj->testRectWrapper(*this, rect));
-}
-
-QRectF
-MyObject::testRectf(const QRectF& rectf)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QRectF, QRectF>{}(
-    m_rustObj->testRectfWrapper(*this, rectf));
-}
-
-QSize
-MyObject::testSize(const QSize& size)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QSize, QSize>{}(
-    m_rustObj->testSizeWrapper(*this, size));
-}
-
-QSizeF
-MyObject::testSizef(const QSizeF& sizef)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QSizeF, QSizeF>{}(
-    m_rustObj->testSizefWrapper(*this, sizef));
-}
-
-QString
-MyObject::testString(const QString& string)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QString, ::std::unique_ptr<QString>>{}(
-    m_rustObj->testStringWrapper(*this, string));
-}
-
-QTime
-MyObject::testTime(const QTime& time)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QTime, QTime>{}(
-    m_rustObj->testTimeWrapper(*this, time));
-}
-
-QUrl
-MyObject::testUrl(const QUrl& url)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QUrl, ::std::unique_ptr<QUrl>>{}(
-    m_rustObj->testUrlWrapper(*this, url));
-}
-
-QVariant
-MyObject::testVariant(const QVariant& variant)
-{
-  const std::lock_guard<std::mutex> guard(*m_rustObjMutex);
-  return rust::cxxqtlib1::cxx_qt_convert<QVariant,
-                                         ::std::unique_ptr<QVariant>>{}(
-    m_rustObj->testVariantWrapper(*this, variant));
-}
-
-} // namespace cxx_qt::my_object
-
-namespace cxx_qt::my_object::cxx_qt_my_object {
-std::unique_ptr<MyObject>
-newCppObject()
-{
-  return std::make_unique<MyObject>();
-}
-} // namespace cxx_qt::my_object::cxx_qt_my_object
