--- crates/cxx-qt-gen/test_outputs/types_qt_invokable.h.license
@@ -1,5 +1,0 @@
-SPDX-FileCopyrightText: 2021 Klarälvdalens Datakonsult AB, a KDAB Group company <info@kdab.com>
-SPDX-FileContributor: Andrew Hayzen <andrew.hayzen@kdab.com>
-SPDX-FileContributor: Gerhard de Clercq <gerhard.declercq@kdab.com>
-
-SPDX-License-Identifier: MIT OR Apache-2.0
